requirements.txt
selenium
applitools
pandas
scikit-learn
google-generativeai
python-dotenv
rich